mcq_video_pipeline/requirements.txt
gspread
google-auth
pandas
gtts
moviepy
elevenlabs
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
